Fix all remaining compilation errors (21 total errors fixed)
This commit fixes all remaining compilation errors found in comprehensive review:

BeanDetailView (7 errors fixed):
- Removed if let for daysFromRoast (Int, not Optional)
- Changed roastLevel/process/variety/tastingNotes from if let to !isEmpty check
- Removed if let for roastDate (Date, not Optional)
- Removed if let for startTime in lastSession

BeansView (1 error fixed):
- Removed if let for daysFromRoast (Int, not Optional)

EditBeanView (5 errors fixed):
- Removed ?? operators for roastLevel, roastDate, process, variety, tastingNotes
- All properties are non-optional in SwiftData Bean model

EditMachineView (3 errors fixed):
- Removed ?? operators for model, boilerType, groupHeadType
- All properties are non-optional in SwiftData Machine model

EditGrinderView (1 error fixed):
- Removed ?? operator for burrType (non-optional String)

FinishBrewView (1 error fixed):
- Changed rating from Int16(rating) to rating
- BrewingSession model expects Int, not Int16

HistoryView (1 error fixed):
- Removed ?? operator for startTime filter (Date, not Optional)

BrewingViewModel (restored):
- Added missing BrewingViewModel.swift from git history
- Required by BrewingView and FinishBrewView

Total: 21 compilation errors resolved across 8 files
All SwiftData non-optional property issues fixed

+++ EspressoTracker/Views/Brewing/FinishBrewView.swift
@@ -218,7 +218,7 @@ struct FinishBrewView: View {
             brewTime: viewModel.elapsedTime,
             waterTemp: waterTemp,
             pressure: pressure,
-            rating: Int16(rating),
+            rating: rating,
             notes: notes,
             imageData: imageData
         )
